% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{as_meltt}
\alias{as_meltt}
\title{as_meltt}
\usage{
as_meltt(.data, ...)
}
\arguments{
\item{.data}{A bundled tbl_df.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
bundled data with additional "meltt_neighbor" column class feature
}
\description{
Coerce a data.frame or tibble object to be class meltt_tbl
}
