% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_close.R
\name{as_neighbor}
\alias{as_neighbor}
\title{as_neighbor}
\usage{
as_neighbor(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt.tbl_df.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
bundled data with additional "meltt_neighbor" column class feature
}
\description{
as_neighbor
}
