% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_taxonomy.R
\name{as_taxonomy}
\alias{as_taxonomy}
\title{as_taxonomy}
\usage{
as_taxonomy(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt.tbl_df.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
bundled data with additional "meltt_assumption" column class feature
}
\description{
as_taxonomy
}
