% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\title{bundle}
\usage{
bundle(...)
}
\arguments{
\item{...}{One or more unquoted expressions (specifically, `data.frame` and `tibble` objects) separated by commas.}
}
\value{
A nested tibble object
}
\description{
Bundle together input data into a nested tibble object for processing in the tidymeltt.
}
\examples{
dat1 = tibble(date=c(1,2,3,4,5))
dat2 = tibble(date=c(1,2,3,4))
bundle(dat1,dat2)


}
