% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_close.R
\name{class_scan}
\alias{class_scan}
\title{class_scan}
\usage{
class_scan(.data, locate_class = NULL)
}
\arguments{
\item{.data}{A bundled meltt_tbl.}

\item{locate_class}{detect it a specific class exists. Default is NULL.}
}
\value{
If locate_class is NULL, returns a list of all classes in each
  bundled column; else a logical value is return if the class exists.
}
\description{
Scan through and report back all classes of the bundled data
}
