% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_close.R
\name{define_close}
\alias{define_close}
\title{define_close}
\usage{
define_close(.data, space, time)
}
\arguments{
\item{.data}{A bundled meltt_tbl containing a `spatial` and `temporal`
features.}

\item{space}{integer that defines the spatial window in terms of kilometers
(km). Default  is 1 km.}

\item{time}{integer that defines the temporal window in terms of days.
Default is 1 day.}
}
\value{
bundled meltt_tbl with a `neighbor` feature. `neighbors` index
  proximate events across bundled datasets
}
\description{
Locates all events from one bundled dataset that fall near (i.e. within a
the same space/time window) an event from another bundled dataset.
}
\examples{

# Generate Fake Spatio-Temporal Data
set.seed(123)
N = 100
dates = seq(from = as.Date("2007-01-01"),to=as.Date("2007-01-01"),by = "day")
lons = runif(N,0,.1)
lats = runif(N,0,.1)
d1 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d2 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d3 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
#d3$date[1:10] = NA

bundle(d1,d2,d3) \%>\%
define_space(.) \%>\%
define_time(d1="date",d2="date",d3="date") \%>\%
define_close(.,space=2,time=1) -> xx


}
