% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_space.R
\name{define_space}
\alias{define_space}
\title{define_space}
\usage{
define_space(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt.tbl_df.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
returns bundle with defined spatial features.
}
\description{
Define the spatial variable as an sf object. `define_time()` and `define_space()` are used to
establish the parameters that set the spatio-temporal window using `proximate()`.
}
\examples{
set.seed(123)
N = 10
dates = seq(from = as.Date("2007-01-01"),to=as.Date("2007-01-01"),by = "day")
lons = runif(N,0,.1)
lats = runif(N,0,.1)
d1 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d2 = data.frame(date = sample(dates,size = N,replace = T),Latitude = sample(lats,N,T),Longitude = sample(lons,N,T))
d3 = data.frame(date = sample(dates,size = N,replace = T),lat = sample(lats,N,T),lon = sample(lons,N,T))

bundle(d1,d2,d3) \%>\% define_space

}
