% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_taxonomy.R
\name{define_taxonomy}
\alias{define_taxonomy}
\title{define_taxonomy}
\usage{
define_taxonomy(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt_tbl.}

\item{...}{pass taxonomies as data.frame or tibble objects. First column
should correspond with the source (as the bundled data is named) and second
variable should correspond with names variable.}
}
\value{
returns bundle with input taxonomy features.
}
\description{
Define the secondary information to be used to disambiguate between
neighboring entries.


Taxonomies:


 * Generalize mutually occurring variables across bundled datasets.

 * Provide schemas that offer guidance on how mutually occurring variables map
onto one another.

 * Provide flexibility in how mutually occurring variables map by creating
levels that go from granular to broad.


Formatting Requirement: Taxonomy data.frames _must_ be formatted as follows:

1. the first column must correspond with the name of the bundled data and be
entitled source.

2. the second column must correspond with the target taxonomy variable. This
name should be consistent across bundled datasets.

Ex: source  variable  level_granular  ...  level_broad
}
\examples{


# Simulate event data
set.seed(123)
D = simulate_event_data(N=50,
                        known.events = 5,
                        known.episodes = 0,
                        known.episodes.events = 0,
                        n.datasets = 4,
                        time.distort = 1,
                        space.distort = 1,
                        episode.window = 10,
                        min_date = as.Date("2001-01-01"),
                        max_date =as.Date("2001-02-01"),
                        lat.bounds = c(-1,1),
                        lon.bounds =c(-1,1),
                        taxonomies = data.frame(k=c(rep(1,4),rep(2,3),rep(3,4)),
                                                depth = c(10,5,3,2, 13,6,3, 15,10,5,2)))
var_1 = D$user.taxonomies$var_1 \%>\% rename(source=data.source)
var_2 = D$user.taxonomies$var_2 \%>\% rename(source=data.source)
var_3 = D$user.taxonomies$var_3 \%>\% rename(source=data.source)
D1 = D$user.data \%>\% filter(dataset=="D1")
D2 = D$user.data \%>\% filter(dataset=="D2")
D3 = D$user.data \%>\% filter(dataset=="D3")
D4 = D$user.data \%>\% filter(dataset=="D4")


bundle(D1,D2,D3,D4) \%>\%
define_taxonomy(var_1,var_2,var_3) -> .data


}
