% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_time.R
\name{define_time}
\alias{define_time}
\title{define_time}
\usage{
define_time(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt.tbl_df.}

\item{...}{pass the name of each primary date column and the bundled dataset it points to.}
}
\value{
returns bundle with defined date features.
}
\description{
Set the temporal variable. `define_time()` and `define_space()` are used to
establish the parameters that set the spatio-temporal window.
}
\details{
Only individual dates are specified, no episodal windows (i.e. start-end
dates). To deal with episodal events, see `episodal_boot()`
}
\examples{

set.seed(123)
N = 10
dates = seq(from = as.Date("2007-01-01"),to=as.Date("2007-01-01"),by = "day")
lons = runif(N,0,.1)
lats = runif(N,0,.1)
d1 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d2 = data.frame(startdate = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d3 = data.frame(date_start = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))

bundle(d1,d2,d3) \%>\% define_time(d1="date",d2="startdate",d3="date_start")

}
