% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disambiguate.R
\name{disambiguate}
\alias{disambiguate}
\title{disambiguate}
\usage{
disambiguate(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt.tbl_df.}

\item{...}{One or more unquoted expressions (specifically, `data.frame` and
`tibble` objects) separated by commas. Data objects are intended to
generalize event categories consistent across bundled data to help
determine which events in the same spatio-temporal bin are unique and which
are matches.}
}
\value{
returns disambiguated data.
}
\description{
Disambiguate between proximate events by leveraging meta-data (secondary
variables) that are conceptually and/or analytically similar across each
bundled datasets.
}
\examples{



}
