% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_date.R
\name{guess_date}
\alias{guess_date}
\title{guess_date}
\usage{
guess_date(date_string, default = "mdy")
}
\arguments{
\item{date_string}{character string containing date information}

\item{default}{default date ordering; see `lubridate` documentation regarding
formatting conventions.}
}
\value{
`Date` vector.
}
\description{
Guesses the most likely date format from a specified date string and convert
it to a `Date` class.
}
\examples{

guess_date(c("2001-03-04","2002-30-04","2012-01-01"))

# If pattern is inconsistent, only the most consistent will be returned
guess_date(c("Jan. 1, 2017","Feb. 13, 2018","2005-06-07"))


}
