% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{is_meltt}
\alias{is_meltt}
\title{is_meltt}
\usage{
is_meltt(x)
}
\arguments{
\item{x}{object being evaluated}
}
\value{
logical scaler if evaluation is true.
}
\description{
Test if a data.set or tibble object is of class meltt_tbl.
}
