% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_close.R
\name{is_neighbor}
\alias{is_neighbor}
\title{is_neighbor}
\usage{
is_neighbor(.data)
}
\arguments{
\item{.data}{A bundled meltt_tbl.}
}
\value{
logical scalar value.
}
\description{
Deterimine if columns in bundled data is a neighbor feature.
}
\examples{

is_neighbor(.data)

}
