% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_taxonomy.R
\name{is_taxonomy}
\alias{is_taxonomy}
\title{is_taxonomy}
\usage{
is_taxonomy(.data)
}
\arguments{
\item{.data}{A bundled meltt_tbl.}
}
\value{
logical scalar value.
}
\description{
Deterimine if columns in bundled data are input assumptions.
}
\examples{

is_taxonomy(.data)

}
