% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_date.R
\name{join_regex}
\alias{join_regex}
\title{Constructs a regex union based on a vector of values}
\usage{
join_regex(...)
}
\description{
Constructs a regex union based on a vector of values
}
\keyword{internal}
