% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_space.R
\name{missing_space}
\alias{missing_space}
\title{missing_space}
\usage{
missing_space(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt.tbl_df.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
prints report of the number/proportion of missing coordinate entries.
}
\description{
Report the number and proportion of missing geo-coordinate entries by
bundeled datasets. Missingness is usually due to an entry being marked as
`NA` or defined as the wrong class.
}
\examples{

set.seed(123)
N = 10
dates = seq(from = as.Date("2007-01-01"),to=as.Date("2007-01-01"),by = "day")
lons = runif(N,0,.1)
lats = runif(N,0,.1)
d1 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d2 = data.frame(date = sample(dates,size = N,replace = T),Latitude = sample(lats,N,T),Longitude = sample(lons,N,T))
d3 = data.frame(date = sample(dates,size = N,replace = T),LATITUDE = sample(lats,N,T),LONGITUDE = sample(lons,N,T))

d3$LATITUDE[1:3] = NA

bundle(d1,d2,d3) \%>\% define_space \%>\% missing_space

}
