% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbundle.R
\name{scrub_meltt_class}
\alias{scrub_meltt_class}
\title{scrub_meltt_class}
\usage{
scrub_meltt_class(.data)
}
\arguments{
\item{.data}{meltt_tbl}

\item{...}{}
}
\description{
drops any meltt_ class features from object. Necessary for unnesting.
}
