% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_date.R
\name{tokens_matched}
\alias{tokens_matched}
\title{Returns strftime tokens already matched in a given string}
\usage{
tokens_matched(string)
}
\description{
Returns strftime tokens already matched in a given string
}
