% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbundle.R
\name{unbundle}
\alias{unbundle}
\title{unbundle}
\usage{
unbundle(.data, ...)
}
\arguments{
\item{.data}{A bundled meltt_tbl containing a `neighbor` feature.}

\item{...}{column requested columns. Default all columns are returned.}
}
\value{
returns tbl_df of all data.
}
\description{
Unbundle (flatten) a bundled meltt_df.
}
\examples{

set.seed(123)
N = 100
dates = seq(from = as.Date("2007-01-01"),to=as.Date("2007-01-01"),by = "day")
lons = runif(N,0,.1)
lats = runif(N,0,.1)
d1 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d2 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
d3 = data.frame(date = sample(dates,size = N,replace = T),latitude = sample(lats,N,T),longitude = sample(lons,N,T))
#d3$date[1:10] = NA

bundle(d1,d2,d3) \%>\%
define_space(.) \%>\%
define_time(d1="date",d2="date",d3="date") \%>\%
define_close(.,space=2,time=1) -> xx


# Unbundle and look at cluster communities
xx \%>\% unbundle(neighbors) \%>\%
group_by(neighbor_id) \%>\% count \%>\%
arrange(desc(n))


}
